Share RichText.Content with native
