+++ packages/block-editor/src/components/index.js
@@ -84,8 +84,8 @@ export {
 	RichTextShortcut,
 	RichTextToolbarButton,
 	__unstableRichTextInputEvent,
-	getRichTextValues as __unstableGetRichTextValues,
 } from './rich-text';
+export { getRichTextValues as __unstableGetRichTextValues } from './rich-text/content';
 export { default as ToolSelector } from './tool-selector';
 export { default as __experimentalUnitControl } from './unit-control';
 export { default as URLInput } from './url-input';
